requirements.txt
streamlit
google-generativeai
python-dotenv
requests
beautifulsoup4
yfinance
finnhub-python
pandas
